+++ packages/sdk/src/subgraph/subgraph.ts
@@ -1,3 +1,6 @@
+import { GraphQLClient } from "graphql-request";
+import { Evt } from "evt";
+import { ChainReader } from "@connext/nxtp-txservice";
 import {
   createLoggingContext,
   FallbackSubgraph,
@@ -5,12 +8,10 @@ import {
   Logger,
   NxtpError,
   RequestContext,
+  SubgraphDomain,
   TransactionData,
   VariantTransactionData,
 } from "@connext/nxtp-utils";
-import { GraphQLClient } from "graphql-request";
-import { Evt } from "evt";
-import { ChainReader } from "@connext/nxtp-txservice";
 
 import { InvalidTxStatus, PollingNotActive } from "../error";
 import {
@@ -123,14 +124,12 @@ export class Subgraph {
     this.chainConfig = {};
     Object.entries(_chainConfig).forEach(([chainId, { subgraph, subgraphSyncBuffer: _subgraphSyncBuffer }]) => {
       const cId = parseInt(chainId);
-      const urls = typeof subgraph === "string" ? [subgraph] : subgraph;
-      const sdksWithClients = urls.map((url) => {
-        return { client: getSdk(new GraphQLClient(url)), url };
-      });
       const fallbackSubgraph = new FallbackSubgraph<Sdk>(
         cId,
-        sdksWithClients,
+        (url: string) => getSdk(new GraphQLClient(url)),
         _subgraphSyncBuffer ?? DEFAULT_SUBGRAPH_SYNC_BUFFER,
+        SubgraphDomain.COMMON,
+        typeof subgraph === "string" ? [subgraph] : subgraph,
       );
       this.sdks[cId] = fallbackSubgraph;
       this.syncStatus[cId] = {
@@ -168,25 +167,16 @@ export class Subgraph {
     }
   }
 
-  getSyncStatus(chainId: number): SubgraphSyncRecord {
+  getSyncStatus(chainId: number): SubgraphSyncRecord | undefined {
     const { requestContext, methodContext } = createLoggingContext(this.getSyncStatus.name);
     const records = this.sdks[chainId].records;
-    this.logger.debug(
-      `Getting sync status for chain ${chainId} with ${records.length} records`,
-      requestContext,
-      methodContext,
-      {
-        chainId,
-        records,
-      },
-    );
-    return (
-      records[0] ?? {
-        synced: false,
-        syncedBlock: 0,
-        latestBlock: 0,
-      }
-    );
+    this.logger.debug("Retrieved subgraph sync status.", requestContext, methodContext, {
+      chainId,
+      records,
+      // For easy searching of "no records found" in logs.
+      recordsFound: records.length,
+    });
+    return records[0];
   }
 
   /**
@@ -526,7 +516,7 @@ export class Subgraph {
     );
 
     // Check to see if errors occurred for all chains (i.e. no active transactions were retrieved).
-    if (errors.size === Object.keys(this.sdks).length) {
+    if (errors.size > 0 && errors.size === Object.keys(this.sdks).length) {
       throw new NxtpError("Failed to get active transactions for all chains due to errors", { errors });
     }
 
@@ -712,13 +702,17 @@ export class Subgraph {
       Object.keys(this.sdks).map(async (_chainId) => {
         const chainId = parseInt(_chainId);
         const subgraph = this.sdks[chainId];
-        const latestBlock = await this.chainReader.getBlockNumber(chainId);
-        const records = await subgraph.sync(latestBlock);
-        const mostSynced = records.sort((r) => r.latestBlock - r.syncedBlock)[0];
+        const records = await subgraph.sync(() => this.chainReader.getBlockNumber(chainId));
+        const mostSynced = records.length > 0 ? records.sort((r) => r.latestBlock - r.syncedBlock)[0] : undefined;
+        this.logger.info("Got most synced.", undefined, undefined, {
+          subgraph,
+          records,
+          mostSynced: typeof mostSynced === "undefined" ? "undefined" : mostSynced,
+        });
         this.syncStatus[chainId] = {
-          latestBlock,
-          syncedBlock: mostSynced.syncedBlock,
-          synced: mostSynced.synced,
+          latestBlock: records[0]?.latestBlock,
+          syncedBlock: mostSynced?.syncedBlock ?? -1,
+          synced: mostSynced?.synced ?? true,
         };
       }),
     );