+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next Release
 
+- [ops] Alpine version update
+- [utils] Implement health endpoint call in FallbackSubgraph
 - [router] Improve logging, add waiting for safe confirmations when handling receiver cancelled
 
 ## 0.1.17